- Fixed some warnings flagged by GCC 4.1.1, which brings up an interesting   observation. The following code works with VC++ 2005. It aborts with GCC:     FString str = "This is a string.";     Printf ("%s\n", str);
SVN r408 (trunk)

+++ src/r_polymost.cpp
@@ -1197,7 +1197,7 @@ void RP_AddLine (seg_t *line)
 	}
 	else
 	{ // The seg is only part of the wall.
-		if (line->linedef->sidenum[0] != line->sidedef - sides)
+		if (line->linedef->sidenum[0] != DWORD(line->sidedef - sides))
 		{
 			swap (v1, v2);
 		}
